requirements.txt
openai-whisper
rich
sounddevice
numpy
torch
tqdm